Initial tweaks for modules create
* Moved custom YAML dumper into nf_core.utils
* Removed default username and made it require an input
* Fixed rich confirm prompt
* Some other minor tidying and bugfixes
